assessment_ms/requirements.txt
Django>=2.2.13,<2.3.0
mysql-connector-python>=8.0.20,<8.1.0
django_ace>=1.0.7,<1.1.0
# sqlalchemy>=1.3.18,<1.4.0
django-simple-history>=2.10.0,<2.11.0